refactor: move get_lightning_loggers to loggers config
